tracing: close plugin with background context
Using the context.Context from the plugin's InitContext as the context
for shutting down the OpenTelemetry tracing SDK can be problematic as
flushing any buffered traces gets skipped if the context is done before
the plugin is Close()d. The v2 shim kicks off the process shutdown
operations by canceling the same context plugins are initialized with,
so the tracing plugin discards any buffered traces on shim shutdown if
the tracing plugin is compiled into the shim. Shut down the
OpenTelemetry SDK using a context derived from the background context to
to prevent the state of the context used to initialize the plugin from
interfering with graceful plugin shutdown.

There is no need to explicitly shut down all the SpanProcessors
ourselves as (*TracerProvider).Shutdown() handles shutting down all
registered span processors itself.

Signed-off-by: Cory Snider <csnider@mirantis.com>